analytics/requirements.txt
Flask==2.0.3
protobuf==3.15.6
google-cloud-datastore